pyproject.toml
[project]
name = "github-statistics-dataset-builder"
version = "0.1.0"
description = "Extract commit statistics from GitHub repositories into Polars-compatible CSV datasets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.0",
    "polars>=1.0.0",
    "typer>=0.12.0",
]

[project.scripts]
github-stats = "github_stats.cli:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/github_stats"]
